major refactoring

+++ src/utils.c
@@ -7,7 +7,7 @@ read_little_endian(FILE *file, unsigned char BitsPerSample) {
 		printf("Bits per sample provided is not a multiple of 8\n");
 		exit(1);
 	}
-	signed short sample = 0;
+	signed long sample = 0;
 	for (int i=0; i<BitsPerSample/8; i++) {
 		unsigned char c = 0;
 		c = fgetc(file);
